Added Kafka consumer

+++ src/main/java/polytech/adce/geolocation/models/Geolocation.java
@@ -33,7 +33,7 @@ public class Geolocation {
         return latitude;
     }
 
-    public void setLatitude(long latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
 
@@ -41,7 +41,7 @@ public class Geolocation {
         return longitude;
     }
 
-    public void setLongitude(long longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
